fix build error for vs2015 and macos

+++ src/scintilla/lexlib/PropSetSimple.h
@@ -11,19 +11,18 @@
 namespace Scintilla {
 
 class PropSetSimple {
-  void* impl;
-  void Set(const char* keyVal);
-
- public:
-  PropSetSimple();
-  virtual ~PropSetSimple();
-  void Set(const char* key, const char* val, size_t lenKey, size_t lenVal);
-  void SetMultiple(const char*);
-  const char* Get(const char* key) const;
-  int GetExpanded(const char* key, char* result) const;
-  int GetInt(const char* key, int defaultValue = 0) const;
+	void *impl;
+	void Set(const char *keyVal);
+public:
+	PropSetSimple();
+	virtual ~PropSetSimple();
+	void Set(const char *key, const char *val, size_t lenKey, size_t lenVal);
+	void SetMultiple(const char *);
+	const char *Get(const char *key) const;
+	size_t GetExpanded(const char *key, char *result) const;
+	int GetInt(const char *key, int defaultValue=0) const;
 };
 
-}  // namespace Scintilla
+}
 
 #endif
